requirements.txt
django==3.2
graphene-django
psycopg2-binary==2.9.5
Pillow==9.3.0
django-structlog==2.2.0
redis==4.3.4
django-redis==5.2.0
django-graphql-jwt==0.3.0
django-graphql-auth==0.3.16
django-mptt==0.14.0
django-phonenumber-field==7.0.0
phonenumbers==8.13.0
graphene-file-upload==1.3.0
rstr==3.2.0
celery==5.2.7